+++ my-app/src/pages/_app.js
@@ -5,7 +5,7 @@ import { ChakraProvider, extendTheme, Flex, Box } from "@chakra-ui/react";
 import NavBar from "../components/NavBar";
 import { Web3Provider } from "@/contexts/Web3Context";
 import { DataBaseProvider } from "@/contexts/DataBaseContext";
-import { TaskBoardProvider } from "@/contexts/TaskBoardContext";
+import { LeaderboardProvider } from "@/contexts/leaderboardContext";
 
 const theme = extendTheme({
   config: {
@@ -24,20 +24,22 @@ function MyApp({ Component, pageProps }) {
   return (
     <DataBaseProvider>
       <Web3Provider>
-        <ChakraProvider theme={theme}>
-          <Flex 
-            direction="column" 
-            minH="100vh" 
-            bg={router.pathname === "/tasks" ? "clear" : "#1285ff"} // Set the background based on the current route
-          >
-            <Box as="header">
-              <NavBar />
-            </Box>
-            <Flex as="main" direction="column" flex="1" overflow="hidden">
-              <Component {...pageProps} />
+        <LeaderboardProvider>
+          <ChakraProvider theme={theme}>
+            <Flex 
+              direction="column" 
+              minH="100vh" 
+              bg={router.pathname === "/tasks" ? "clear" : "#1285ff"} // Set the background based on the current route
+            >
+              <Box as="header">
+                <NavBar />
+              </Box>
+              <Flex as="main" direction="column" flex="1" overflow="hidden">
+                <Component {...pageProps} />
+              </Flex>
             </Flex>
-          </Flex>
-        </ChakraProvider>
+          </ChakraProvider>
+        </LeaderboardProvider>
       </Web3Provider>
     </DataBaseProvider>
   );
